+++ R/backend.R
@@ -7,6 +7,12 @@
 #' saved into googlesheet
 #' @param output_ss Character vector with output googlesheet id (if \code{output_gsheet = T})
 #' @param output_sheet Character vector with output spreadsheet name (if \code{output_gsheet = T})
+#' @param div_id Character vector with div id
+#' @param css parsed css line to include for the survey
+#' @param button_labels Character vector with labels for both active and inactive buttons
+#' @param render_ui reactiveVal holding the logical indicating if the UI should be
+#' rendered or not
+#' @param module_ui_id Character vector with id for UI elements that will be generated
 #'
 #' @import shiny
 #' @import shinyjs
@@ -19,11 +25,26 @@
   output_gsheet,
   output_ss,
   output_sheet,
-  button_labels
+  div_id,
+  css,
+  button_labels,
+  render_ui,
+  module_ui_id
 ){
   moduleServer(
     id = id,
     function(input, output, session) {
+
+      output$quetzio_UI <- renderUI({
+
+        req(render_ui())
+
+        .generate_ui(source_list = source_list,
+                     div_id = div_id,
+                     css = css,
+                     button_label = button_labels[1],
+                     module_ui_id = module_ui_id)}
+      )
 
       # reactiveValues for storing valid and mandatory inputs status
       # as valid$mandatory_filled and valid$minmax_matched
@@ -41,6 +62,7 @@
 
       observe({
 
+        req(!is.null(input$submit))
         valid$mandatory_ids <- names(source_list)[mandatory_items]
         valid$numeric_ids <- names(source_list)[numeric_items]
 
@@ -146,17 +168,17 @@
             updateActionButton(session,
                                inputId = "submit",
                                label = "Submitted!",
-                               icon = icon("fa-thumbs-up"))
+                               icon = icon("thumbs-up"))
           },
           error = function(err){
 
-            status$is_done <- as.logical(NA)
+            status$is_done <- FALSE
             status$message <- err
 
             updateActionButton(session,
                                inputId = "submit",
                                label = "Error occured",
-                               icon = icon("fa-frown-open"))
+                               icon = icon("frown-open"))
 
           }
           )